requirements.txt
Django==2.1.1
djangorestframework==3.8.2
django-rest-swagger==2.2.0
markdown==2.1.0
coreapi==1.32.0
django-filter==1.0.1
